requirements.txt
flask
peewee
pillow
